src/main/java/com/example/jpabook/repository/MemberBasicJPARepository.java
package com.example.jpabook.repository;

import com.example.jpabook.domain.Member;
import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class MemberBasicJPARepository {

    @PersistenceContext
    private EntityManager em;

    public Long save(Member member){
        em.persist(member);
        return member.getId();
    }

    public Member findOne(Long id){

        return em.find(Member.class, id);
    }

}
